Add files via upload
This script procedurally generates a specified number of objects across a specified-size area. 
Make sure to add the ObjectBuilderEditor script to the Editor folder.
Then Add the ObjectBuilderScript to an empty game object, and drag a prefab of an object that you want to spawn to the "Object to Spawn" property field.